refactor: modularize adventure feature
