+++ scripts/validate-structure.js
@@ -148,7 +148,7 @@ class StructureValidator {
     const coreFiles = [
       'src/game/state.js',
       'src/game/engine.js',
-      'src/game/adventure.js',
+      'src/features/adventure/logic.js',
       'ui/index.js',
       'index.html',
       'docs/ai-verification-protocol.md'
