+++ src/features/adventure/mutators.js
@@ -1,0 +1,10 @@
+export {
+  selectZone,
+  selectArea,
+  selectAreaById,
+  startAdventureCombat,
+  startBossCombat,
+  progressToNextArea,
+  retreatFromCombat,
+  instakillCurrentEnemy
+} from './logic.js';
