+++ src/features/adventure/state.js
@@ -1,0 +1,20 @@
+export const adventureState = {
+  currentZone: 0,
+  currentArea: 0,
+  selectedZone: 0,
+  selectedArea: 0,
+  totalKills: 0,
+  areasCompleted: 0,
+  zonesUnlocked: 1,
+  killsInCurrentArea: 0,
+  bestiary: {},
+  inCombat: false,
+  areaProgress: {},
+  unlockedAreas: { "0-0": true },
+  currentEnemy: null,
+  enemyHP: 0,
+  enemyMaxHP: 0,
+  playerHP: 0,
+  combatLog: [],
+  sessionLoot: []
+};
